update prepare.py to remove dependency of .ssh/id_rsa
on line 192 --- this path is only relevant if the user is using ssh keys. If he is using  ssh via password then this line breaks plumbery. 
We need to add logic so this should only be checked in the case where ssh keys are being used for authentication.
Even then should this be hardcoded?

Unfortunately Iam not sure what authentication model the user is using otherwise I would have submitted actualy code isntead of a comment